after food delete and quantity set

+++ src/Pages/My Orders/OrderList.jsx
@@ -1,46 +1,5 @@
-// import React, { use } from "react";
-// import OrderTable from "./OrderTable";
-
-// const OrderList = ({ myOrderPromise }) => {
-//   const orders = use(myOrderPromise);
-
-//   return (
-//     <div>
-//       <h3>orders done : {orders.length}</h3>
-//       <div className="overflow-x-auto my-10 w-3/5 mx-auto">
-//         <table className="table">
-//           {/* head */}
-//           <thead>
-//             <tr>
-//               <th>
-//                 #
-//               </th>
-//               <th>Food Name</th>
-//               <th>Food Price</th>
-//               <th>Quantity</th>
-//               <th>Buying Date</th>
-//               <th>Buyer Name</th>
-//               <th>Remove Items</th>
-//             </tr>
-//           </thead>
-//           <tbody>
-//             {orders.map((order, index) => (
-//               <OrderTable
-//                 order={order}
-//                 index={index + 1}
-//                 key={order._id}
-//               ></OrderTable>
-//             ))}
-//           </tbody>
-//         </table>
-//       </div>
-//     </div>
-//   );
-// };
-
-// export default OrderList;
-
-import React, { use, useState } from "react";
+// import React, { use, useState } from "react";
+import { use, useState } from "react";
 import OrderTable from "./OrderTable";
 import axios from "axios";
 
@@ -52,10 +11,26 @@ const OrderList = ({ myOrderPromise }) => {
     const confirm = window.confirm("Are you sure you want to cancel this order?");
     if (!confirm) return;
 
+    // Find the order to delete (to get foodId and quantity)
+    const orderToDelete = orders.find(order => order._id === id);
+    if (!orderToDelete) return;
+
     axios.delete(`http://localhost:3000/purchase/${id}`)
       .then(res => {
         if (res.data.deletedCount > 0) {
+          // Remove from UI
           setOrders(prev => prev.filter(order => order._id !== id));
+
+          // Restore quantity in menu
+          axios.patch(`http://localhost:3000/foods/${orderToDelete.foodId}/restore`, {
+            quantity: orderToDelete.quantity
+          })
+          .then(() => {
+            console.log("Quantity restored in menu.");
+          })
+          .catch(err => {
+            console.error("Failed to restore quantity:", err);
+          });
         }
       })
       .catch(err => {
